runtime: Refactor, corrected mismatched coordinate data types.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@1575 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUISliderControl.cpp
@@ -91,12 +91,12 @@ void CGUISliderControl::Render()
 	{
 		if (m_bHasFocus)
 		{
-			m_guiMidFocus.SetPosition((int)fPos, m_guiBackground.GetYPosition() );
+			m_guiMidFocus.SetPosition((int)fPos, (int)m_guiBackground.GetYPosition() );
 			m_guiMidFocus.Render();
 		}
 		else
 		{
-			m_guiMid.SetPosition((int)fPos, m_guiBackground.GetYPosition() );
+			m_guiMid.SetPosition((int)fPos, (int)m_guiBackground.GetYPosition() );
 			m_guiMid.Render();
 		}
 	}
